+++ src/homematicip/connection/connection_url_resolver.py
@@ -72,7 +72,4 @@ class ConnectionUrlResolver:
     def _get_verify(enforce_ssl: bool, ssl_context):
         if ssl_context is not None:
             return ssl_context
-        if enforce_ssl:
-            return enforce_ssl
-
-        return True
+        return enforce_ssl